backend/requirements.txt
django
djangorestframework
pandas
numpy
scikit-learn
joblib
reportlab
django-cors-headers
gunicorn